KernelSU: Conditionally hardcode KSU_VERSION for clarity
By default, KernelSU is configured to iterate versioning based on
number of revisions.

This becomes misleading in our case as making a change to the
submodule for compatibility with this kernel results in KSU_VERSION
incrementing from 11928 > 11929.

This implies that we are running a version of KSU that we actually
aren't, as version 11929 from the official repo is representative
of the latest stable tag +1 revision which would be divergent from
our local "11929".

In order to make it clearer that we are using a "custom" variant of
KernelSU, hardcode KSU_VERSION to reflect the current stable
release and add a tag to indicate that it has been modified for use
with this kernel.

Retain the default versioning logic if !CONFIG_LOCALVERSION="-NeutrinoKernel".

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>

[EmanuelCN: Modified to actually hardcode the ksu version no matter the kernel string]